Refactor Datasets tab components into separate (presentational and container) component files
