+++ static/js/components/Delete/index.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
+
 const Delete = (props) => {
   const style = {
     display: 'inline-block'
