+++ static/js/components/Main.jsx
@@ -3,7 +3,6 @@ import PropTypes from 'prop-types';
 import { connect, Provider } from 'react-redux';
 import ReactDOM from 'react-dom';
 import ReactTabs from 'react-tabs';
-import PropTypes from 'prop-types';
 
 import 'bootstrap-css';
 import 'bootstrap';
@@ -12,16 +11,14 @@ import { Notifications } from 'baselayer/components/Notifications';
 import WebSocket from 'baselayer/components/WebSocket';
 import configureStore from '../configureStore';
 import * as Action from '../actions';
-import WebSocket from 'baselayer/components/WebSocket';
 import CesiumMessageHandler from '../CesiumMessageHandler';
 import AddProject from '../containers/AddProject';
 import ProjectSelector from '../containers/ProjectSelector';
 import ProjectTab from '../containers/ProjectTab';
-import DatasetsTab from './Datasets';
+import DatasetsTab from './DatasetsTab';
 import FeaturesTab from './Features';
 import ModelsTab from './Models';
 import PredictTab from './Predictions';
-import { Notifications } from 'baselayer/components/Notifications';
 import colorScheme from './colorscheme';
 import Progress from './Progress';
 import CesiumTooltip from './Tooltip';
