--- static/js/containers/DatasetForm/DatasetForm.jsx/index.jsx
@@ -1,34 +1,0 @@
-import reduxForm from 'redux-form';
-
-import * as Validate from '../../validate';
-import * as Action from '../../actions';
-import DatasetForm from '../../containers/DatasetForm';
-
-
-const mapStateToProps = (state, ownProps) => (
-  {
-    initialValues: {
-      ...(ownProps.initialValues),
-      projectID: ownProps.selectedProject.id
-    }
-  }
-);
-
-const mapDispatchToProps = (dispatch, ownProps) => (
-  {
-    onSubmit: form => (
-      dispatch(Action.uploadDataset(form))
-    )
-  }
-);
-
-const validate = Validate.createValidator({
-  datasetName: [Validate.required],
-  tarFile: [Validate.oneFile],
-});
-
-export default reduxForm({
-  form: 'newDataset',
-  fields: ['datasetName', 'headerFile', 'tarFile', 'projectID'],
-  validate
-}, dsMapStateToProps, dsMapDispatchToProps)(DatasetForm);
